# specs/openapi/testdata-factory.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TestData Factory API",
    "version": "0.1.0",
    "description": "Self-hosted API for contract validation and deterministic test data generation."
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "API is healthy"
          }
        }
      }
    },
    "/v1/model-profiles": {
      "get": {
        "summary": "List local model profiles",
        "responses": {
          "200": {
            "description": "Supported model profiles"
          }
        }
      }
    },
    "/v1/contracts/validate": {
      "post": {
        "summary": "Validate a contract",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Structured contract validation feedback",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "422": {
            "description": "Request payload is malformed"
          }
        }
      }
    },
    "/v1/data/generate": {
      "post": {
        "summary": "Generate deterministic test data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated test data"
          },
          "422": {
            "description": "Generation failed"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ContractPayload": {
        "type": "object",
        "required": ["contract"],
        "properties": {
          "contract": {
            "type": "object"
          }
        }
      },
      "GeneratePayload": {
        "type": "object",
        "required": ["contract", "scenarioId"],
        "properties": {
          "contract": {
            "type": "object"
          },
          "scenarioId": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "minimum": 1,
            "default": 1
          },
          "seed": {
            "type": "string"
          }
        }
      },
      "ValidationFinding": {
        "type": "object",
        "required": ["severity", "field", "message", "recommendation"],
        "properties": {
          "severity": {
            "type": "string",
            "enum": ["info", "warning", "error"]
          },
          "field": {
            "type": ["string", "null"],
            "description": "Contract field or path associated with the finding, when available."
          },
          "message": {
            "type": "string"
          },
          "recommendation": {
            "type": "string"
          }
        }
      },
      "ValidationResult": {
        "type": "object",
        "required": ["status", "score", "findings"],
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["valid", "needs_review", "invalid"]
          },
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationFinding"
            }
          }
        }
      }
    }
  }
}